+++ tencentcloud/partners/v20180321/models.go
@@ -659,7 +659,7 @@ type DescribeAgentAuditedClientsRequest struct {
 	// 限制数目
 	Limit *uint64 `json:"Limit,omitempty" name:"Limit"`
 
-	// 客户类型：可以为new(新拓)/assign(指定)/old(存量已关联)/old_newchecking(存量-新关联考核中)/old_newnotpass(存量-新关联未达标)/direct(直销)/direct_newopp(直销(新商机))/空
+	// 可以为new(自拓)/assign(指派)/old(官网)/direct(直销)/direct_newopp(直销(新商机))/空
 	ClientType *string `json:"ClientType,omitempty" name:"ClientType"`
 
 	// 项目类型：可以为self(自拓项目)/platform(合作项目)/repeat(复算项目  )/空