+++ tencentcloud/vod/v20180717/models.go
@@ -2023,8 +2023,9 @@ type AudioTemplateInfo struct {
 	// <li>libmp3lame：更适合 flv；</li>
 	// <li>mp2。</li>
 	// 当外层参数 Container 为 hls 时，可选值为：
-	// <li>libfdk_aac；</li>
-	// <li>libmp3lame。</li>
+	// <li>libfdk_aac。</li>
+	// 当外层参数 Format 为 HLS 或 MPEG-DASH 时，可选值为：
+	// <li>libfdk_aac。</li>
 	Codec *string `json:"Codec,omitempty" name:"Codec"`
 
 	// 音频流的码率，取值范围：0 和 [26, 256]，单位：kbps。
@@ -2063,8 +2064,9 @@ type AudioTemplateInfoForUpdate struct {
 	// <li>libmp3lame：更适合 flv；</li>
 	// <li>mp2。</li>
 	// 当外层参数 Container 为 hls 时，可选值为：
-	// <li>libfdk_aac；</li>
-	// <li>libmp3lame。</li>
+	// <li>libfdk_aac。</li>
+	// 当外层参数 Format 为 HLS 或 MPEG-DASH 时，可选值为：
+	// <li>libfdk_aac。</li>
 	Codec *string `json:"Codec,omitempty" name:"Codec"`
 
 	// 音频流的码率，取值范围：0 和 [26, 256]，单位：kbps。 当取值为 0，表示音频码率和原始音频保持一致。
@@ -2859,7 +2861,8 @@ type CreateAdaptiveDynamicStreamingTemplateRequest struct {
 	*tchttp.BaseRequest
 
 	// 自适应转码格式，取值范围：
-	// <li>HLS。</li>
+	// <li>HLS；</li>
+	// <li>MPEG-DASH。</li>
 	Format *string `json:"Format,omitempty" name:"Format"`
 
 	// 自适应转码输出子流参数信息，最多输出10路子流。
@@ -10441,11 +10444,15 @@ type ModifyAdaptiveDynamicStreamingTemplateRequest struct {
 	// 自适应转码模板唯一标识。
 	Definition *uint64 `json:"Definition,omitempty" name:"Definition"`
 
+	// <b>点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。</b>
+	SubAppId *uint64 `json:"SubAppId,omitempty" name:"SubAppId"`
+
 	// 模板名称，长度限制：64 个字符。
 	Name *string `json:"Name,omitempty" name:"Name"`
 
 	// 自适应转码格式，取值范围：
-	// <li>HLS。</li>
+	// <li>HLS；</li>
+	// <li>MPEG-DASH。</li>
 	Format *string `json:"Format,omitempty" name:"Format"`
 
 	// 是否禁止视频低码率转高码率，取值范围：
@@ -10464,9 +10471,6 @@ type ModifyAdaptiveDynamicStreamingTemplateRequest struct {
 
 	// 模板描述信息，长度限制：256 个字符。
 	Comment *string `json:"Comment,omitempty" name:"Comment"`
-
-	// 点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。
-	SubAppId *uint64 `json:"SubAppId,omitempty" name:"SubAppId"`
 }
 
 func (r *ModifyAdaptiveDynamicStreamingTemplateRequest) ToJsonString() string {
@@ -10482,13 +10486,13 @@ func (r *ModifyAdaptiveDynamicStreamingTemplateRequest) FromJsonString(s string)
 		return err
 	}
 	delete(f, "Definition")
+	delete(f, "SubAppId")
 	delete(f, "Name")
 	delete(f, "Format")
 	delete(f, "DisableHigherVideoBitrate")
 	delete(f, "DisableHigherVideoResolution")
 	delete(f, "StreamInfos")
 	delete(f, "Comment")
-	delete(f, "SubAppId")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "ModifyAdaptiveDynamicStreamingTemplateRequest has unknown keys!", "")
 	}